+++ src/msg.c
@@ -92,7 +92,7 @@ void MSG_Copy(msg_t *buf, byte *data, int length, msg_t *src)
 =============================================================================
 
 bit functions
-  
+
 =============================================================================
 */
 
@@ -124,7 +124,7 @@ void MSG_WriteBits( msg_t *msg, int value, int bits ) {
 		else if(bits==16)
 		{
 			short temp = value;
-			
+
 			CopyLittleShort(&msg->data[msg->cursize], &temp);
 			msg->cursize += 2;
 			msg->bit += 16;
@@ -135,7 +135,7 @@ void MSG_WriteBits( msg_t *msg, int value, int bits ) {
 			msg->cursize += 4;
 			msg->bit += 32;
 		}
-		else 
+		else
 			return; //Com_Error(ERR_DROP, "can't write %d bits", bits);
 	} else {
 //		fp = fopen("c:\\netchan.bin", "a");
@@ -187,7 +187,7 @@ int MSG_ReadBits( msg_t *msg, int bits ) {
 		else if(bits==16)
 		{
 			short temp;
-			
+
 			CopyLittleShort(&temp, &msg->data[msg->readcount]);
 			value = temp;
 			msg->readcount += 2;
@@ -354,22 +354,22 @@ void MSG_WriteAngle16( msg_t *sb, float f ) {
 // returns -1 if no more characters are available
 int MSG_ReadChar (msg_t *msg ) {
 	int	c;
-	
+
 	c = (signed char)MSG_ReadBits( msg, 8 );
 	if ( msg->readcount > msg->cursize ) {
 		c = -1;
-	}	
-	
+	}
+
 	return c;
 }
 
 int MSG_ReadByte( msg_t *msg ) {
 	int	c;
-	
+
 	c = (unsigned char)MSG_ReadBits( msg, 8 );
 	if ( msg->readcount > msg->cursize ) {
 		c = -1;
-	}	
+	}
 	return c;
 }
 
@@ -386,41 +386,41 @@ int MSG_LookaheadByte( msg_t *msg ) {
 
 int MSG_ReadShort( msg_t *msg ) {
 	int	c;
-	
+
 	c = (short)MSG_ReadBits( msg, 16 );
 	if ( msg->readcount > msg->cursize ) {
 		c = -1;
-	}	
+	}
 
 	return c;
 }
 
 int MSG_ReadLong( msg_t *msg ) {
 	int	c;
-	
+
 	c = MSG_ReadBits( msg, 32 );
 	if ( msg->readcount > msg->cursize ) {
 		c = -1;
-	}	
-	
+	}
+
 	return c;
 }
 
 float MSG_ReadFloat( msg_t *msg ) {
 	floatint_t dat;
-	
+
 	dat.i = MSG_ReadBits( msg, 32 );
 	if ( msg->readcount > msg->cursize ) {
 		dat.f = -1;
-	}	
-	
-	return dat.f;	
+	}
+
+	return dat.f;
 }
 
 char *MSG_ReadString( msg_t *msg ) {
 	static char	string[MAX_STRING_CHARS];
 	int		l,c;
-	
+
 	l = 0;
 	do {
 		c = MSG_ReadByte(msg);		// use ReadByte so -1 is out of bounds
@@ -438,17 +438,17 @@ char *MSG_ReadString( msg_t *msg ) {
 
 		string[l] = c;
 		l++;
-	} while (l < sizeof(string)-1);
-	
+	} while (l < (int) sizeof(string)-1);
+
 	string[l] = 0;
-	
+
 	return string;
 }
 
 char *MSG_ReadBigString( msg_t *msg ) {
 	static char	string[BIG_INFO_STRING];
 	int		l,c;
-	
+
 	l = 0;
 	do {
 		c = MSG_ReadByte(msg);		// use ReadByte so -1 is out of bounds
@@ -466,10 +466,10 @@ char *MSG_ReadBigString( msg_t *msg ) {
 
 		string[l] = c;
 		l++;
-	} while (l < sizeof(string)-1);
-	
+	} while (l < (int) sizeof(string)-1);
+
 	string[l] = 0;
-	
+
 	return string;
 }
 
@@ -494,10 +494,10 @@ char *MSG_ReadStringLine( msg_t *msg ) {
 
 		string[l] = c;
 		l++;
-	} while (l < sizeof(string)-1);
-	
+	} while (l < (int) sizeof(string)-1);
+
 	string[l] = 0;
-	
+
 	return string;
 }
 
@@ -537,7 +537,7 @@ int MSG_HashKey(const char *string, int maxlen) {
 =============================================================================
 
 delta functions
-  
+
 =============================================================================
 */
 
@@ -585,7 +585,7 @@ float MSG_ReadDeltaFloat( msg_t *msg, float oldV ) {
 =============================================================================
 
 delta functions with keys
-  
+
 =============================================================================
 */
